relocate workplan creation fixtures for reuse
